+++ domain/pricing.go
@@ -112,8 +112,7 @@ type PricingConfig struct {
 	MaxRoutes        int `mapstructure:"max-routes"`
 	// MinPoolLiquidityCap is the minimum liquidity capitalization required for a pool to be considered in the router.
 	MinPoolLiquidityCap uint64 `mapstructure:"min-pool-liquidity-cap"`
-
-	// WorkerMinPoolLiquidityCap is the minimum liquidity capitalization required for a pool to be considered in the pricing worker.
+	// WorkerMinPoolLiquiidtyCap is the minimum liquidity capitalization required for a pool to be considered in the pricing worker.
 	WorkerMinPoolLiquidityCap uint64 `mapstructure:"worker-min-pool-liquidity-cap"`
 }
 
@@ -153,14 +152,23 @@ type PoolLiquidityPricerWorker interface {
 	// RepriceDenomMetadata reprices the token liquidity metadata for the denoms updated within the block.
 	// Returns the updated token metadata.
 	// If there is an update for a denom with a later height than the current height, it is skipped, making this a no-op.
+	// If denom is a gamm share, it is also skipped.
 	// Relies on the blockPriceUpdates to get the price for the denoms.
 	// If the price for denom cannot be fetched, the liquidity capitalization for this denom is set to zero.
 	// The latest update height for this denom is updated on completion.
-	RepriceDenomMetadata(updateHeight uint64, blockPriceUpdates PricesResult, quoteDenom string, blockDenomLiquidityUpdatesMap BlockPoolMetadata) PoolDenomMetaDataMap
+	RepriceDenomsMetadata(updateHeight uint64, blockPriceUpdates PricesResult, quoteDenom string, blockDenomLiquidityUpdatesMap BlockPoolMetadata) PoolDenomMetaDataMap
+	// CreatePoolDenomMetaData creates a pool denom metatata by finding the total liquidity across all pools in block pool metadata,
+	// retrieving the price from blockPriceUpdates and recomputing the liquidity capitalization for the given denom, update height, block price updates and quote denom.
+	// Returns the pool denom metadata and error if any.
+	// Returns error if:
+	// - the denom pool liquidity data is not found for the updatedBlockDenom.
+	// - the price is not found for the given denom.
+	// Note that in case of an error, the pool denom metadata is still returned with values
+	// computed as best-effort (set to zero if cannot compute).
+	CreatePoolDenomMetaData(updatedBlockDenom string, updateHeight uint64, blockPriceUpdates PricesResult, quoteDenom string, blockPoolMetadata BlockPoolMetadata) (PoolDenomMetaData, error)
 
-	// GetLatestUpdateHeightForDenom returns the latest height for when the liquidity for a given denom was updated
-	// zero if the height is not found or fails to cast it to the return type.
-	GetLatestUpdateHeightForDenom(denom string) uint64
+	// GetHeightForDenom returns zero if the height is not found or fails to cast it to the return type.
+	GetHeightForDenom(denom string) uint64
 
 	// StoreHeightForDenom stores the latest height for the given denom.
 	StoreHeightForDenom(denom string, height uint64)
@@ -220,5 +228,5 @@ func (prices PricesResult) GetPriceForDenom(baseDenom string, quoteDenom string)
 		return osmomath.ZeroBigDec()
 	}
 
-	return price
+	return price.Clone()
 }